+++ README.md
@@ -24,6 +24,11 @@ To install all hooks from Conan repository in Github:
 
 ``$ conan config install https://github.com/conan-io/hooks``
 
+If you are using Conan >=1.14 you can specify the source and destination folder to avoid copying
+undesired files to your local cache:
+
+``$ conan config install https://github.com/conan-io/hooks -sf hooks -tf hooks ``
+
 Conan config install does not activate any hook.
 
 ## Hook activation
